assorted lang fixes

+++ examples/objects/dxWriter/dxWrite.cc
@@ -5,11 +5,11 @@
 int main()
 {
     Rappture::DXWriter d;
-    float i = 0.0;
+    double i = 0.0;
     int rank = 3;
     size_t p[] = {3,5,5};
-    float origin[] = {0.0,0.0,0.0};
-    float delta[] = {1.0,0.0,0.0,0.0,1.0,0.0,0.0,0.0,1.0};
+    double origin[] = {0.0,0.0,0.0};
+    double delta[] = {1.0,0.0,0.0,0.0,1.0,0.0,0.0,0.0,1.0};
 
     d.rank(rank);
     d.origin(origin);
